% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attacking_direction.R
\name{add_attacking_direction}
\alias{add_attacking_direction}
\title{Add the attacking direction to events}
\usage{
add_attacking_direction(events, tDAT, metaDF, GameCentre)
}
\arguments{
\item{events}{An events dataframe created from the 'parse_f24' function}

\item{tDAT}{A tracking dataframe created from the 'parse_tracking' function}

\item{metaDF}{An tracking metadata dataframe created from the 'parse_CH_metadata' function}

\item{GameCentre}{A list of dataframes created from the 'parse_f7' function}
}
\value{
An events dataframe with the added attacking.direction variable
}
\description{
Adding a variable of 'attacking.direction' to the events dataframe allows us to keep
track of the direction of play and how that relates to x, y positioning. It's a basic
but crucial variable to add to our events dataframe to improve our analysis
}
