% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CH_origin.R
\name{add_CH_origin}
\alias{add_CH_origin}
\title{Convert the pass origin from OPTA to ChyronHego spec}
\usage{
add_CH_origin(df)
}
\arguments{
\item{df}{An events dataframe created from the 'parse_event' function}
}
\value{
An events dataframe with the added origin.x and origin.y variables
}
\description{
OPTA event data has two variables x and y, that represent the the x and y pitch
coordinates for the origina point of an event. We can convert from OPTA 0-100 values to 
ChyronHego CM and return the events dataframe back with new variables of origin.x and 
origin.y. The attacking.direction needs to be added to the events dataframe first from 
the 'add_attacking_direction' function in this package.
}
