% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pass_target.R
\name{add_pass_target}
\alias{add_pass_target}
\title{Convert the pass destination from OPTA to ChyronHego spec}
\usage{
add_pass_target(df)
}
\arguments{
\item{df}{An events dataframe created from the 'parse_event' function}
}
\value{
An events dataframe with the added target.x and target.y variables
}
\description{
OPTA event data has two variables 140 and 141, that represent the the x and y pitch
coordinates for the end point of a pass. We can convert from OPTA 0-100 values to 
ChyronHego CM and return the events dataframe back with new variables of target.x and 
target.y. The attacking.direction needs to be added to the events dataframe first from 
the 'add_attacking_direction' function in this package.
}
