% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_TimeSync.R
\name{add_TimeSync}
\alias{add_TimeSync}
\title{Add a TimeSync variable}
\usage{
add_TimeSync(events, metaDF)
}
\arguments{
\item{events}{An events dataframe created from the 'parse_f24' function}

\item{metaDF}{An tracking metadata dataframe created from the 'parse_CH_metadata' function}
}
\value{
An events dataframe with the added TimeSync variable
}
\description{
The OPTA and ChyronHego datasets don't share a common clock. OPTA data is recorded in seconds and ChyronHego
data is recorded at 25 frames a second. The addition of a TimeSync variable to the events dataframe allows 
the two dataframes to be synced and linked together easier with a common clock.
}
