% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_OPTA_pitch.R
\name{create_OPTA_pitch}
\alias{create_OPTA_pitch}
\title{Create a Pitch ready for OPTA data}
\usage{
create_OPTA_pitch(grass_colour = "#F9F9F9", line_colour = "#8F8F8F",
  background_colour = "#F9F9F9", goal_colour = "#000000",
  goaltype = "barcanumbers", middlethird = TRUE, BasicFeatures = FALSE,
  arcs = TRUE, padding = 5)
}
\arguments{
\item{grass_colour}{Colour of the grass, Default = "#F9F9F9", HEXCODE or accept colour string i.e. "red"}

\item{line_colour}{Colour of the line, Default = "#8F8F8F", HEXCODE or accept colour string i.e. "red"}

\item{background_colour}{Colour of the Default = "#F9F9F9", border area, HEXCODE or accept colour string i.e. "red"}

\item{goaltype}{Type of goal used. Default = "barcanumbers". "line" and "box" also available}

\item{middlethird}{Default = FALSE, Adds a shading to the middle third of the pitch}

\item{BasicFeatures}{Default = FALSE, If TRUE removes most of the pitch features to have a more minimalist design}

\item{arcs}{Default = TRUE, adds the D-arcs}

\item{padding}{Default = 10, adds some padding around the pitch}
}
\value{
A pitch plot, that allows data to be plotted on top
}
\description{
A pitch plot base that is designed for the OPTA values for x and y of 0-100
Uses ggplot and returns a plot ready for further data to be plotted over the top.
}
