% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_CH_metadata.R
\name{parse_CH_metadata}
\alias{parse_CH_metadata}
\title{Parse an ChyronHego Metadata File}
\usage{
parse_CH_metadata(xml.filename)
}
\arguments{
\item{xml.filename}{Path to the input file}
}
\value{
A dataframe of the Tracking metadata
}
\description{
A ChyronHego Metadata file contains key information in order to use the actual 
Tracking data. This simple function takes the path to an xml file and returns a
a parsed dataframe.
}
