% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_f24.R
\name{parse_f24}
\alias{parse_f24}
\title{Parse an OPTA f24 Files}
\usage{
parse_f24(xml.filename)
}
\arguments{
\item{xml.filename}{Path to the input file}
}
\value{
A dataframe of the OPTA event data
}
\description{
An OPTA f24 file is an xml file with all of the data for each event of a match.
This function takes the file location and parses the XML file into a dataframe. 
All qualifiers are in columns named after them, access them using ``. If the
qualifier doesn't have a value but was present it is represented by the value 1.
}
