% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_f24_multiple.R
\name{parse_f24_multiple}
\alias{parse_f24_multiple}
\title{Parse Multiple OPTA f24 Files}
\usage{
parse_f24_multiple(xml.filename)
}
\arguments{
\item{xml.filename}{Path to the input file}
}
\value{
A dataframe of the OPTA event data
}
\description{
An OPTA f24 file is an xml file with all of the data for each event of a match.
This function takes the file locations and parses the XML files into a dataframe. 
All qualifiers are in columns named after them, access them using ``. If the
qualifier doesn't have a value but was present it is represented by the value 1.
Make sure the XML file is named like '/999999_f24.xml' and the location is named 
in such a way so that there are no numbers other than match ID and f24, the 
file location will be used to calculate the match ID.
}
