% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_f7.R
\name{parse_f7}
\alias{parse_f7}
\title{Parse an OPTA f7 Files}
\usage{
parse_f7(f7.xml, MatchID)
}
\arguments{
\item{f7.xml}{Path to the input file}

\item{MatchID}{The ID number of the game to be parsed}
}
\value{
A list of dataframes containing match information
}
\description{
An OPTA f7 file is an xml file with various information about a match such as 
lineups, goals, cards, kits used and more. This function takes the file location 
and parses the XML file into a dataframes which are stored in a list.
}
