% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tracking.R
\name{parse_tracking}
\alias{parse_tracking}
\title{Parse an ChyronHego Tracking Data File}
\usage{
parse_tracking(tracking.xml, metadata)
}
\arguments{
\item{tracking.xml}{Path to the input file}

\item{metadata}{Dataframe of metadata as created via parse_CH_metadata() in this package}
}
\value{
A dataframe of the parsed tracking data
}
\description{
ChyronHego's tracking data contains positional data at 25 frames per second.
The raw data comes in .dat format with all players, officials and ball positional data
per frame combined in a string. This function spilts up the string and arranges the 
resulting data in a logical format.
}
